[IMP] web: convert dialog test to hoot
task-4028335
